fixed AMD loading
fixed swichwgt default false states

had some success with parsley

Was breaking when using RequireJS. So changed Grunt Build so that Validator.js returns an object with Assert and Validator properties and self-executing wrapper function stores it in a global var called 'Validator'. Also changed ui.js so that excludes are not affected by errorsContainer.